pyproject.toml
[tool.poetry]
name = "animelist"
version = "0.1.0"
description = ""
authors = ["Watashi <watashi@localhost>"]

[tool.poetry.dependencies]
python = "^3.8"
pycairo = "^1.20.0"
PyGObject = "^3.38.0"
SQLAlchemy = "^1.3.22"
python-decouple = "^3.3"
SQLAlchemy-Utils = "^0.36.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
